requirements.txt
flask==2.0.1
flask-cors==3.0.10
pandas==1.3.3
numpy==1.21.2
scikit-learn==0.24.2
nltk==3.6.2
textblob==0.15.3
python-dotenv==0.19.0
requests==2.26.0
google-api-python-client==2.31.0
python-dateutil==2.8.2
matplotlib==3.4.3
seaborn==0.11.2
werkzeug==2.0.1
scipy>=1.11.0
transformers==4.38.2
spacy==3.7.2
plotly==5.19.0
reportlab==4.1.0
sentence-transformers==2.5.1 